requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
opencv-python
Pillow
tqdm
gdown
